Remove Tailwind CSS and Integrate Mantine UI
The codebase is transitioned from Tailwind CSS to Mantine UI as per needs for more sophisticated UI components which Mantine provides out of the box. Tailwind related configuration and dependencies have been removed. Necessary Mantine dependencies are added. Emotion is also added for more advanced CSS-in-JS capabilities. MantineProvider is integrated in _app. Changes are also made in .eslintrc and prettier.config to ensure linting and formatting works with the new setup.