Add layout, header, and new pages; enhance login function, routing and more
This commit includes several enhancements, starting from adding a layout and header with a mobile-responsive design to the frontend. It also added tabs for navigation. There are new pages for workouts, exercises, and the user profile. Additionally, the login function now includes options for both sign-in and sign-out. The routing was also improved. Lastly, '404 - Page Not Found' page was added and various parts of our website code were updated to adhere to best coding practices. The color scheme of the UI has been changed to dark mode. These changes dramatically improve the usability and appearance of our application.