Enhanced Exercise View and Data Retrieval Process
Added functionality to read and display detailed exercise data from the database, including type, muscle category, and group. Improved former UI, made it more user-friendly by adding selection labels. Implemented new form validation for the Exercise data model and provided CRUD operations for Exercise data on the server side. Updated the prisma schema to include a reference to the User who created an Exercise. This was done to extend the functionality of the application and improve user experience.