Added user session to exercise view
Added a useSession hook to the exercise page, which allows us to match the user session with the exercise creator id. This aims to provide edit and delete functionality only to the respective exercise creators, enhancing the security and integrity of user-generated content. Also, changed a label from "Total Sessions" to "Total Workouts" for better clarity.